+++ src/main/java/org/opensearch/knn/plugin/transport/KNNProfileTransportAction.java
@@ -5,7 +5,6 @@
 
 package org.opensearch.knn.plugin.transport;
 
-import org.apache.commons.math3.stat.descriptive.StatisticalSummaryValues;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.opensearch.action.support.ActionFilters;
@@ -22,22 +21,20 @@ import org.opensearch.core.action.support.DefaultShardOperationFailedException;
 import org.opensearch.core.common.io.stream.StreamInput;
 import org.opensearch.indices.IndicesService;
 import org.opensearch.knn.index.KNNIndexShard;
-import org.opensearch.transport.TransportService;
 import org.opensearch.threadpool.ThreadPool;
+import org.opensearch.transport.TransportService;
 
 import java.io.IOException;
 import java.util.List;
 
-/**
- * Transport action for profiling KNN vectors in an index
- */
 public class KNNProfileTransportAction extends TransportBroadcastByNodeAction<
     KNNProfileRequest,
     KNNProfileResponse,
-    KNNProfileShardResult> {
+    KNNIndexShardProfileResult> {
 
     public static Logger logger = LogManager.getLogger(KNNProfileTransportAction.class);
-    private final IndicesService indicesService;
+
+    private IndicesService indicesService;
 
     @Inject
     public KNNProfileTransportAction(
@@ -60,8 +57,8 @@ public class KNNProfileTransportAction extends TransportBroadcastByNodeAction<
     }
 
     @Override
-    protected KNNProfileShardResult readShardResult(StreamInput in) throws IOException {
-        return new KNNProfileShardResult(in);
+    protected KNNIndexShardProfileResult readShardResult(StreamInput in) throws IOException {
+        return new KNNIndexShardProfileResult(null, null);
     }
 
     @Override
@@ -70,11 +67,11 @@ public class KNNProfileTransportAction extends TransportBroadcastByNodeAction<
         int totalShards,
         int successfulShards,
         int failedShards,
-        List<KNNProfileShardResult> shardResults,
+        List<KNNIndexShardProfileResult> profileResults,
         List<DefaultShardOperationFailedException> shardFailures,
         ClusterState clusterState
     ) {
-        return new KNNProfileResponse(totalShards, successfulShards, failedShards, shardResults, shardFailures);
+        return new KNNProfileResponse(profileResults, totalShards, successfulShards, failedShards, shardFailures);
     }
 
     @Override
@@ -83,19 +80,12 @@ public class KNNProfileTransportAction extends TransportBroadcastByNodeAction<
     }
 
     @Override
-    protected KNNProfileShardResult shardOperation(KNNProfileRequest request, ShardRouting shardRouting) throws IOException {
+    protected KNNIndexShardProfileResult shardOperation(KNNProfileRequest request, ShardRouting shardRouting) throws IOException {
         KNNIndexShard knnIndexShard = new KNNIndexShard(
             indicesService.indexServiceSafe(shardRouting.shardId().getIndex()).getShard(shardRouting.shardId().id())
         );
 
-        List<StatisticalSummaryValues> profileResults = knnIndexShard.profile(request.getFieldName());
-        logger.info(
-            "[KNN] Profile completed for field: {} on shard: {} - stats count: {}",
-            request.getFieldName(),
-            shardRouting.shardId(),
-            profileResults != null ? profileResults.size() : 0
-        );
-        return new KNNProfileShardResult(shardRouting.shardId(), profileResults);
+        return knnIndexShard.profile(request.getField());
     }
 
     @Override